+++ enginetest/sqllogictest/convert/convert.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/dolthub/go-mysql-server/enginetest/tmp/utils"
+	"github.com/dolthub/go-mysql-server/enginetest/sqllogictest/utils"
 
 	_ "github.com/go-sql-driver/mysql"
 )
